tests: Move ErrorMonitor into RenderFramework
For error reporting at DestroyInstance-time, the ErrorMonitor
instantiation must outlive the VkLayerTest class. Moved it into
the VkRenderFramework class, which is the parent of VkLayerTest.

Fixes crashes when using the provided debug callback at
DestroyInstance time.
