+++ tests/layer_validation_tests.h
@@ -223,75 +223,6 @@ T NearestSmaller(const T from) {
     return std::nextafter(from, negative_direction);
 }
 
-// ErrorMonitor Usage:
-//
-// Call SetDesiredFailureMsg with a string to be compared against all
-// encountered log messages, or a validation error enum identifying
-// desired error message. Passing NULL or VALIDATION_ERROR_MAX_ENUM
-// will match all log messages. logMsg will return true for skipCall
-// only if msg is matched or NULL.
-//
-// Call VerifyFound to determine if all desired failure messages
-// were encountered. Call VerifyNotFound to determine if any unexpected
-// failure was encountered.
-class ErrorMonitor {
-   public:
-    ErrorMonitor();
-
-    ~ErrorMonitor();
-
-    // Set monitor to pristine state
-    void Reset();
-
-    // ErrorMonitor will look for an error message containing the specified string(s)
-    void SetDesiredFailureMsg(const VkFlags msgFlags, const std::string msg);
-    void SetDesiredFailureMsg(const VkFlags msgFlags, const char *const msgString);
-
-    // ErrorMonitor will look for an error message containing the specified string(s)
-    template <typename Iter>
-    void SetDesiredFailureMsg(const VkFlags msgFlags, Iter iter, const Iter end) {
-        for (; iter != end; ++iter) {
-            SetDesiredFailureMsg(msgFlags, *iter);
-        }
-    }
-
-    // Set an error that the error monitor will ignore. Do not use this function if you are creating a new test.
-    // TODO: This is stopgap to block new unexpected errors from being introduced. The long-term goal is to remove the use of this
-    // function and its definition.
-    void SetUnexpectedError(const char *const msg);
-
-    VkBool32 CheckForDesiredMsg(const char *const msgString);
-    vector<string> GetOtherFailureMsgs() const;
-    VkDebugReportFlagsEXT GetMessageFlags() const;
-    bool AnyDesiredMsgFound() const;
-    bool AllDesiredMsgsFound() const;
-    void SetError(const char *const errorString);
-    void SetBailout(bool *bailout);
-    void DumpFailureMsgs() const;
-
-    // Helpers
-
-    // ExpectSuccess now takes an optional argument allowing a custom combination of debug flags
-    void ExpectSuccess(VkDebugReportFlagsEXT const message_flag_mask = VK_DEBUG_REPORT_ERROR_BIT_EXT);
-
-    void VerifyFound();
-    void VerifyNotFound();
-
-   private:
-    // TODO: This is stopgap to block new unexpected errors from being introduced. The long-term goal is to remove the use of this
-    // function and its definition.
-    bool IgnoreMessage(std::string const &msg) const;
-
-    VkFlags message_flags_;
-    std::unordered_multiset<std::string> desired_message_strings_;
-    std::unordered_multiset<std::string> failure_message_strings_;
-    std::vector<std::string> ignore_message_strings_;
-    vector<string> other_messages_;
-    test_platform_thread_mutex mutex_;
-    bool *bailout_;
-    bool message_found_;
-};
-
 class VkLayerTest : public VkRenderFramework {
    public:
     void VKTriangleTest(BsoFailSelect failCase);
@@ -303,11 +234,9 @@ class VkLayerTest : public VkRenderFramework {
               const VkCommandPoolCreateFlags flags = 0, void *instance_pnext = nullptr);
     bool AddSurfaceInstanceExtension();
     bool AddSwapchainDeviceExtension();
-    ErrorMonitor *Monitor();
     VkCommandBufferObj *CommandBuffer();
 
    protected:
-    ErrorMonitor *m_errorMonitor;
     uint32_t m_instance_api_version = 0;
     uint32_t m_target_api_version = 0;
     bool m_enableWSI;
@@ -319,7 +248,6 @@ class VkLayerTest : public VkRenderFramework {
         PFN_vkGetOriginalPhysicalDeviceFormatPropertiesEXT &fpvkGetOriginalPhysicalDeviceFormatPropertiesEXT);
 
     VkLayerTest();
-    ~VkLayerTest();
 };
 
 class VkPositiveLayerTest : public VkLayerTest {
